requirements.txt
grpcio==1.44.0
grpcio-tools==1.44.0
numpy==1.22.3
paho-mqtt==1.6.1
protobuf==3.19.4
six==1.16.0
